login successful

+++ ftp-server/src/main/java/com/nazmul/ftp/server/protocol/ProtocolCode.java
@@ -2,11 +2,11 @@ package com.nazmul.ftp.server.protocol;
 
 public enum ProtocolCode {
     ;
-    public static final int RRQ = 100;
-    public static final int WRQ = 200;
-    public static final int DATA = 300;
-    public static final int ACK = 400;
-    public static final int ERROR = 500;
-    public static final int LOGIN = 600;
-    public static final int LOGOUT = 700;
+    public static final short RRQ = 100;
+    public static final short WRQ = 200;
+    public static final short DATA = 300;
+    public static final short ACK = 400;
+    public static final short ERROR = 500;
+    public static final short LOGIN = 600;
+    public static final short LOGOUT = 700;
 }
